Regenerate translation files
There have been a couple of string rearrangements and some new string
addition with the GTK3 port. Some of them (menu labels) are
automatically detected by intltool and should translate fine. Button
tooltips are marked as fuzzy, it should be safe to just accept the
proposed translation, at least it is so with the italian translation but
we may need translators input for this.